+++ utils/build.base.js
@@ -7,21 +7,24 @@ function build(channel, callback) {
   var webpack = require("webpack"),
     path = require("path"),
     fs = require("fs"),
-    config = require("../webpack.config"),
-    ZipPlugin = require("zip-webpack-plugin")
+    config = require("../webpack.config")
 
   delete config.chromeExtensionBoilerplate
 
   config.mode = "production"
 
-  var packageInfo = JSON.parse(fs.readFileSync("package.json", "utf-8"))
+  // Only add zip plugin if PACKAGE env var is set
+  if (process.env.PACKAGE === "true") {
+    var ZipPlugin = require("zip-webpack-plugin")
+    var packageInfo = JSON.parse(fs.readFileSync("package.json", "utf-8"))
 
-  config.plugins = (config.plugins || []).concat(
-    new ZipPlugin({
-      filename: `${packageInfo.name}-${packageInfo.version}-${channel}.zip`,
-      path: path.join(__dirname, "../", "zip")
-    })
-  )
+    config.plugins = (config.plugins || []).concat(
+      new ZipPlugin({
+        filename: `${packageInfo.name}-${packageInfo.version}-${channel}.zip`,
+        path: path.join(__dirname, "../", "zip")
+      })
+    )
+  }
 
   webpack(config, function (err, stats) {
     if (err || stats.hasErrors()) {